+++ Yunbeom/src/test/java/com/depromeet/yunbeom/post/controller/PostCreateControllerTest.java
@@ -1,25 +1,20 @@
-package com.depromeet.yunbeom.controller;
+package com.depromeet.yunbeom.post.controller;
 
-import static org.mockito.ArgumentMatchers.*;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;
 
 import org.junit.jupiter.api.Test;
-import org.mockito.BDDMockito;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.http.MediaType;
-import org.springframework.mail.SimpleMailMessage;
-import org.springframework.mail.javamail.JavaMailSender;
 import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.context.jdbc.Sql;
 import org.springframework.test.context.jdbc.SqlGroup;
 import org.springframework.test.web.servlet.MockMvc;
 
-import com.depromeet.yunbeom.user.domain.UserCreate;
+import com.depromeet.yunbeom.post.domain.PostCreate;
 import com.fasterxml.jackson.databind.ObjectMapper;
 
 @SpringBootTest
@@ -27,39 +22,38 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 @AutoConfigureMockMvc
 @AutoConfigureTestDatabase
 @SqlGroup({
+	// 테스트 실행하기 전에 실행합니다.
+	@Sql(value = "/sql/post-controller-test-data.sql", executionPhase = Sql.ExecutionPhase.BEFORE_TEST_METHOD),
 	// 테스트 실행한 후에 실행합니다.
 	@Sql(value = "/sql/delete-all-data.sql", executionPhase = Sql.ExecutionPhase.AFTER_TEST_METHOD)
 })
-class UserCreateControllerTest {
+class PostCreateControllerTest {
 
 	@Autowired
 	private MockMvc mockMvc;
-	@MockBean
-	private JavaMailSender mailSender;
 	@Autowired
 	private ObjectMapper objectMapper;
 
 	@Test
-	void 사용자는_회원_가입을_할_수있고_회원가입된_사용자는_PENDING_상태이다() throws Exception {
+	void 사용자는_게시물을_작성할_수_있다() throws Exception {
 		// given
-		UserCreate userCreate = UserCreate.builder()
-			.email("ybchar@kakao.com")
-			.nickname("ybchar")
-			.address("Pangyo")
+		PostCreate postCreate = PostCreate.builder()
+			.writerId(1)
+			.content("helloworld")
 			.build();
-		BDDMockito.doNothing().when(mailSender).send(any(SimpleMailMessage.class));
 
 		// when
 		// then
 		mockMvc.perform(
-				post("/api/users")
+				post("/api/posts")
 					.contentType(MediaType.APPLICATION_JSON)
-					.content(objectMapper.writeValueAsString(userCreate)))
+					.content(objectMapper.writeValueAsString(postCreate)))
 			.andExpect(status().isCreated())
 			.andExpect(jsonPath("$.id").isNumber())
-			.andExpect(jsonPath("$.email").value("ybchar@kakao.com"))
-			.andExpect(jsonPath("$.nickname").value("ybchar"))
-			.andExpect(jsonPath("$.status").value("PENDING"));
+			.andExpect(jsonPath("$.content").value("helloworld"))
+			.andExpect(jsonPath("$.writer.id").isNumber())
+			.andExpect(jsonPath("$.writer.email").value("uiurihappy@naver.com"))
+			.andExpect(jsonPath("$.writer.nickname").value("uiurihappy"));
 	}
 
 }
