+++ Yunbeom/src/test/java/com/depromeet/yunbeom/user/controller/UserControllerTest.java
@@ -1,4 +1,4 @@
-package com.depromeet.yunbeom.controller;
+package com.depromeet.yunbeom.user.controller;
 
 import static org.assertj.core.api.AssertionsForClassTypes.*;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*;
